+++ src/main/java/kahlua/KahluaProject/repository/ticket/TicketCustomRepository.java
@@ -1,4 +1,4 @@
-package kahlua.KahluaProject.repository.ticket.TicketRepository;
+package kahlua.KahluaProject.repository.ticket;
 
 import kahlua.KahluaProject.domain.ticket.Ticket;
 import kahlua.KahluaProject.domain.ticket.Type;
