+++ src/main/java/kahlua/KahluaProject/repository/ticket/TicketRepository.java
@@ -1,4 +1,4 @@
-package kahlua.KahluaProject.repository.ticket.TicketRepository;
+package kahlua.KahluaProject.repository.ticket;
 
 import kahlua.KahluaProject.domain.ticket.Status;
 import kahlua.KahluaProject.domain.ticket.Ticket;
